pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "avengine"
version = "0.1.0"
description = "Autonomous Multi-Agent Video Creation Engine"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }

dependencies = [
    # ── Web Framework ──────────────────────────────────────────────────────────
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "httpx>=0.27.0",
    "websockets>=13.0",
    "python-multipart>=0.0.9",

    # ── Task Queue & Pub/Sub ───────────────────────────────────────────────────
    "celery[redis]>=5.4.0",
    "redis>=5.0.0",
    "psycopg2-binary>=2.9.0",

    # ── Multi-Agent Orchestration ──────────────────────────────────────────────
    "langgraph>=0.2.0",
    "langchain>=0.3.0",
    "langchain-openai>=0.2.0",
    "langchain-anthropic>=0.2.0",
    "crewai>=0.80.0",
    "crewai-tools>=0.14.0",
    "pydantic-ai>=0.0.13",

    # ── Data Validation & Settings ─────────────────────────────────────────────
    "pydantic>=2.9.0",
    "pydantic-settings>=2.5.0",

    # ── Video/Audio Processing ─────────────────────────────────────────────────
    "moviepy>=2.0.0",
    "yt-dlp>=2024.11.0",
    "ffmpeg-python>=0.2.0",

    # ── Browser Automation ─────────────────────────────────────────────────────
    "playwright>=1.48.0",

    # ── Web Scraping ───────────────────────────────────────────────────────────
    "beautifulsoup4>=4.12.0",
    "feedparser>=6.0.11",
    "lxml>=5.3.0",
    "aiohttp>=3.10.0",

    # ── TwelveLabs SDK ─────────────────────────────────────────────────────────
    "twelvelabs>=0.3.0",

    # ── Vector Database ────────────────────────────────────────────────────────
    "qdrant-client>=1.12.0",
    "lancedb>=0.13.0",

    # ── ElevenLabs TTS ─────────────────────────────────────────────────────────
    "elevenlabs>=1.9.0",

    # ── Observability ─────────────────────────────────────────────────────────
    "opentelemetry-api>=1.27.0",
    "opentelemetry-sdk>=1.27.0",
    "opentelemetry-instrumentation-fastapi>=0.48b0",
    "opentelemetry-exporter-otlp>=1.27.0",
    "structlog>=24.4.0",
    "sentry-sdk[fastapi]>=2.16.0",

    # ── Storage & DB ──────────────────────────────────────────────────────────
    "sqlalchemy[asyncio]>=2.0.36",
    "alembic>=1.13.0",
    "asyncpg>=0.30.0",
    "boto3>=1.35.0",
    "aioboto3>=13.1.0",

    # ── Utilities ─────────────────────────────────────────────────────────────
    "tenacity>=9.0.0",
    "python-ulid>=3.0.0",
    "orjson>=3.10.0",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=5.0.0",
    "pytest-mock>=3.14.0",
    "httpx>=0.27.0",
    "ruff>=0.7.0",
    "mypy>=1.13.0",
    "pre-commit>=4.0.0",
    "factory-boy>=3.3.0",
    "respx>=0.21.0",
]

[project.scripts]
avengine = "src.main:main"

[tool.hatch.build.targets.wheel]
packages = ["src"]

# ── Ruff (linting + formatting) ────────────────────────────────────────────────
[tool.ruff]
target-version = "py311"
line-length = 100
src = ["src", "tests"]

[tool.ruff.lint]
select = [
    "E", "W",   # pycodestyle
    "F",        # pyflakes
    "I",        # isort
    "N",        # pep8-naming
    "UP",       # pyupgrade
    "B",        # flake8-bugbear
    "C4",       # flake8-comprehensions
    "SIM",      # flake8-simplify
    "TCH",      # flake8-type-checking
    "ANN",      # flake8-annotations
    "S",        # flake8-bandit (security)
    "RUF",      # ruff-specific
]
ignore = ["ANN101", "ANN102", "S101"]

[tool.ruff.lint.per-file-ignores]
"tests/**" = ["ANN", "S"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

# ── MyPy ──────────────────────────────────────────────────────────────────────
[tool.mypy]
python_version = "3.11"
strict = true
ignore_missing_imports = true
plugins = ["pydantic.mypy"]

# ── Pytest ────────────────────────────────────────────────────────────────────
[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
addopts = "--cov=src --cov-report=term-missing --cov-fail-under=80"

[tool.coverage.run]
source = ["src"]
omit = ["*/migrations/*", "*/tests/*"]
